# docs/swagger.yaml
definitions:
  dto.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.CalculateCampaignCapacityRequest:
    properties:
      adlink:
        maxLength: 10000
        type: string
      budget:
        type: integer
      city:
        items:
          type: string
        maxItems: 255
        type: array
      content:
        maxLength: 512
        minLength: 1
        type: string
      line_number:
        maxLength: 255
        type: string
      scheduleat:
        type: string
      segment:
        maxLength: 255
        type: string
      sex:
        maxLength: 255
        type: string
      subsegment:
        items:
          type: string
        maxItems: 255
        type: array
      title:
        maxLength: 255
        type: string
    type: object
  dto.CalculateCampaignCapacityResponse:
    properties:
      capacity:
        type: integer
      message:
        type: string
    type: object
  dto.CalculateCampaignCostRequest:
    properties:
      adlink:
        maxLength: 10000
        type: string
      budget:
        type: integer
      city:
        items:
          type: string
        maxItems: 255
        type: array
      content:
        maxLength: 512
        minLength: 1
        type: string
      line_number:
        maxLength: 255
        type: string
      scheduleat:
        type: string
      segment:
        maxLength: 255
        type: string
      sex:
        maxLength: 255
        type: string
      subsegment:
        items:
          type: string
        maxItems: 255
        type: array
      title:
        maxLength: 255
        type: string
    type: object
  dto.CalculateCampaignCostResponse:
    properties:
      max_msg_target:
        type: integer
      message:
        type: string
      msg_target:
        type: integer
      total:
        type: integer
    type: object
...
  dto.CreateCampaignRequest:
    properties:
      adlink:
        maxLength: 10000
        type: string
      budget:
        type: integer
      city:
        items:
          type: string
        maxItems: 255
        type: array
      content:
        maxLength: 512
        minLength: 1
        type: string
      line_number:
        maxLength: 255
        type: string
      schedule_at:
        type: string
      segment:
        maxLength: 255
        type: string
      sex:
        maxLength: 255
        type: string
      subsegment:
        items:
          type: string
        maxItems: 255
        type: array
      title:
        maxLength: 255
        type: string
    type: object
  dto.CreateCampaignResponse:
    properties:
      created_at:
        type: string
      message:
        type: string
      status:
        type: string
      uuid:
        type: string
    type: object
...
  dto.GetCampaignResponse:
    properties:
      adlink:
        type: string
      budget:
        type: integer
      city:
        items:
          type: string
        type: array
      comment:
        type: string
      content:
        type: string
      created_at:
        type: string
      line_number:
        type: string
      scheduleat:
        type: string
      segment:
        type: string
      sex:
        type: string
      status:
        type: string
      subsegment:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
...
  dto.ListCampaignsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.GetCampaignResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/dto.PaginationInfo'
    type: object
...
  dto.PaginationInfo:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.UpdateCampaignRequest:
    properties:
      adlink:
        maxLength: 10000
        type: string
      budget:
        type: integer
      city:
        items:
          type: string
        maxItems: 255
        type: array
      content:
        maxLength: 512
        minLength: 1
        type: string
      finalize:
        type: boolean
      line_number:
        maxLength: 255
        type: string
      scheduleat:
        type: string
      segment:
        maxLength: 255
        type: string
      sex:
        maxLength: 255
        type: string
      subsegment:
        items:
          type: string
        maxItems: 255
        type: array
      title:
        maxLength: 255
        type: string
    type: object
  dto.UpdateCampaignResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/campaigns:
    get:
      consumes:
      - application/json
      description: Retrieve the authenticated user's campaigns with pagination, ordering,
        and filters
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Items per page (max 100)
        in: query
        name: limit
        required: true
        type: integer
      - default: newest
        description: Order by (newest|oldest)
        in: query
        name: orderby
        type: string
      - description: Filter by title (contains)
        in: query
        name: title
        type: string
      - description: Filter by status (initiated|in-progress|waiting-for-approval|approved|rejected)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ListCampaignsResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: List Campaigns
      tags:
      - Campaigns
    post:
      consumes:
      - application/json
      description: Create a new campaign with the specified parameters
      parameters:
      - description: Campaign creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCampaignRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Campaign created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CreateCampaignResponse'
              type: object
        "400":
          description: Validation error or invalid request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized - customer not found or inactive
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Create Campaign
      tags:
      - Campaigns
  /api/v1/campaigns/{uuid}:
    put:
      consumes:
      - application/json
      description: Update an existing campaign with the specified parameters
      parameters:
      - description: Campaign UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Campaign update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCampaignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Campaign updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UpdateCampaignResponse'
              type: object
        "400":
          description: Validation error or invalid request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized - customer not found or inactive
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden - campaign access denied or update not allowed
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Campaign not found
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Update Campaign
      tags:
      - Campaigns
  /api/v1/campaigns/calculate-capacity:
    post:
      consumes:
      - application/json
      description: Calculate the potential reach and capacity of an campaign based
        on parameters
      parameters:
      - description: Campaign parameters for capacity calculation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CalculateCampaignCapacityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Capacity calculated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CalculateCampaignCapacityResponse'
              type: object
        "400":
          description: Validation error or invalid request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Calculate Campaign Capacity
      tags:
      - Campaigns
  /api/v1/campaigns/calculate-cost:
    post:
      consumes:
      - application/json
      description: Calculate the total cost of an campaign based on content, target
        audience, and pricing factors
      parameters:
      - description: Campaign parameters for cost calculation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CalculateCampaignCostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cost calculated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CalculateCampaignCostResponse'
              type: object
        "400":
          description: Validation error or invalid request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Calculate Campaign Cost
      tags:
      - Campaigns
...
swagger: "2.0"